move ActiveProgram and ProgramState
