+++ common/src/benchmarking.rs
@@ -18,7 +18,10 @@
 
 use super::*;
 
-use gear_core::{pages::WasmPage, program::MemoryInfix};
+use gear_core::{
+    pages::{WasmPage, WasmPagesAmount},
+    program::MemoryInfix,
+};
 use gear_wasm_instrument::parity_wasm::{self, elements::*};
 use sp_io::hashing::blake2_256;
 use sp_runtime::traits::Zero;
