+++ core/Cargo.toml
@@ -27,6 +27,7 @@ enum-iterator.workspace = true
 byteorder.workspace = true
 num-traits.workspace = true
 numerated.workspace = true
+primitive-types = { workspace = true, features = ["scale-info"] }
 
 # Optional dependencies
 serde = { workspace = true, features = ["derive"], optional = true }
