+++ pallets/gear/src/lib.rs
@@ -50,7 +50,7 @@ pub use crate::{
     pallet::*,
     schedule::{InstructionWeights, Limits, MemoryWeights, Schedule, SyscallWeights},
 };
-pub use gear_core::{gas::GasInfo, message::ReplyInfo};
+pub use gear_core::{gas::GasInfo, message::ReplyInfo, program::ProgramState};
 pub use weights::WeightInfo;
 
 use alloc::{
@@ -60,7 +60,7 @@ use alloc::{
 use common::{
     self, event::*, gas_provider::GasNodeId, paused_program_storage::SessionId, scheduler::*,
     storage::*, BlockLimiter, CodeMetadata, CodeStorage, GasProvider, GasTree, Origin,
-    PausedProgramStorage, Program, ProgramState, ProgramStorage, QueueRunner,
+    PausedProgramStorage, Program, ProgramStorage, QueueRunner,
 };
 use core::marker::PhantomData;
 use core_processor::{
