+++ pallets/gear/src/runtime_api.rs
@@ -18,14 +18,13 @@
 
 use super::*;
 use crate::queue::QueueStep;
-use common::ActiveProgram;
 use core::convert::TryFrom;
 use frame_support::{dispatch::RawOrigin, traits::PalletInfo};
 use gear_core::{
     code::TryNewCodeConfig,
     message::ReplyInfo,
     pages::{numerated::tree::IntervalsTree, WasmPage},
-    program::MemoryInfix,
+    program::{ActiveProgram, MemoryInfix},
 };
 use gear_wasm_instrument::syscalls::SyscallName;
 use sp_runtime::{DispatchErrorWithPostInfo, ModuleError};
