+++ pallets/gear/src/tests.rs
@@ -32,7 +32,7 @@ use crate::{
     ProgramStorageOf, QueueOf, Schedule, TaskPoolOf, WaitlistOf,
 };
 use common::{
-    event::*, scheduler::*, storage::*, ActiveProgram, CodeStorage, GasTree, LockId, LockableTree,
+    event::*, scheduler::*, storage::*, CodeStorage, GasTree, LockId, LockableTree,
     Origin as _, Program, ProgramStorage, ReservableTree,
 };
 use core_processor::common::ActorExecutionErrorReplyReason;
@@ -50,6 +50,7 @@ use gear_core::{
         ReplyInfo, StoredDispatch, UserStoredMessage,
     },
     pages::WasmPage,
+    program::ActiveProgram,
 };
 use gear_core_backend::error::{
     TrapExplanation, UnrecoverableExecutionError, UnrecoverableExtError, UnrecoverableWaitError,
@@ -7474,7 +7475,7 @@ fn gas_spent_precalculated() {
 
         let get_program_code = |pid| {
             let code_id = ProgramStorageOf::<Test>::get_program(pid)
-                .and_then(|program| common::ActiveProgram::try_from(program).ok())
+                .and_then(|program| ActiveProgram::try_from(program).ok())
                 .expect("program must exist")
                 .code_hash
                 .cast();
@@ -14927,6 +14928,7 @@ pub(crate) mod utils {
         ids::{CodeId, MessageId, ProgramId},
         message::{Message, Payload, ReplyDetails, UserMessage, UserStoredMessage},
         reservation::GasReservationMap,
+        program::ActiveProgram,
     };
     use gear_core_errors::*;
     use pallet_gear_voucher::VoucherId;
@@ -15163,7 +15165,7 @@ pub(crate) mod utils {
             let expected_code = ProgramCodeKind::OutgoingWithValueInHandle.to_bytes();
             assert_eq!(
                 ProgramStorageOf::<Test>::get_program(prog_id)
-                    .and_then(|program| common::ActiveProgram::try_from(program).ok())
+                    .and_then(|program| ActiveProgram::try_from(program).ok())
                     .expect("program must exist")
                     .code_hash,
                 generate_code_hash(&expected_code).into(),
@@ -15183,7 +15185,7 @@ pub(crate) mod utils {
         )
         .into();
         let actual_code_hash = ProgramStorageOf::<Test>::get_program(program_id)
-            .and_then(|program| common::ActiveProgram::try_from(program).ok())
+            .and_then(|program| ActiveProgram::try_from(program).ok())
             .map(|prog| prog.code_hash)
             .expect("invalid program address for the test");
         assert_eq!(
@@ -15608,7 +15610,7 @@ pub(crate) mod utils {
     #[track_caller]
     pub(super) fn get_reservation_map(pid: ProgramId) -> Option<GasReservationMap> {
         let program = ProgramStorageOf::<Test>::get_program(pid).unwrap();
-        if let common::Program::Active(common::ActiveProgram {
+        if let common::Program::Active(ActiveProgram {
             gas_reservation_map,
             ..
         }) = program
